cleaned up dead files, create comp field mapping file
