--- packages/amplify-ui-components/src/components/amplify-password-field/amplify-password-field.style.ts
@@ -1,9 +1,0 @@
-// import { css } from 'emotion';
-
-// export const forgotPasswordLink = css`
-//   font-family: var(--font-family);
-//   font-weight: 300;
-//   font-size: var(--third-font-size);
-//   color: var(--grey);
-//   margin-bottom: 2.625rem;
-// `;
