+++ packages/amplify-ui-components/src/components/amplify-sign-in/amplify-sign-in.tsx
@@ -60,6 +60,7 @@ export class AmplifySignIn {
    * Form fields allows you to utilize our pre-built components such as username field, code field, password field, email field, etc.
    * by passing an array of strings that you would like the order of the form to be in. If you need more customization, such as changing
    * text for a label or adjust a placeholder, you can follow the structure below in order to do just that.
+   * ```
    * [
    *  {
    *    type: 'username'|'password'|'email'|'code'|'default',
@@ -69,6 +70,7 @@ export class AmplifySignIn {
    *    required: boolean
    *  }
    * ]
+   * ```
   */
   @Prop() formFields: object[] | string[] = signInComponents;
